gnss: f_ref trim-bleed re-pin FIXED + LIVE-VALIDATED (phase 1, still default-OFF)
Three bugs found and fixed by the staged live validation (L2C-only arm); the
mechanism now VERIFIES. Still --carrier-bleed default 0.

  1. SIGN. The combined carrier is f_ref - ctrim (the NCO rides the r2c-flipped
     internal convention -- "phi enters NEGATED"), NOT f_ref + ctrim. So the fold is
     f_ref -= trim, not +=. With +=, the loop re-grew the trim to ~2x (measured:
     PRN5 +2.53 -> +5.13) and the sat decohered. With -=, f_ref = (true+ctrim)-ctrim
     = true: VERIFIED, trim -> ~0, coherent.
  2. AMOUNT not flag. carrier_repin is now the absorbed amount (Hz), passed broker ->
     tracker (car_repin_pending: set -> dict), so the tracker knows how far to move
     f_ref. (f_ref = dop, the first attempt, adopted the FLEET model -- off from the
     per-sat carrier by exactly the trim -> also decohered.)
  3. ONE-SHOT consume. The anchor loop runs once PER RECORD in a frame while repin is
     a frame-wide snapshot; clear repin[p] after applying so the fold lands exactly
     once, not N_records times.

Live result (L2C armed, 2026-08-03): re-pin received + applied once; VERIFIED cases
fold a ~5 Hz standing trim into f_ref, trim -> +0.05, sat coherent through the verify
window (phase-continuous, no deep-window reset). Safety machinery held throughout
(REFUTED -> loop re-grows from 0 + 120 s lockout, no fleet collapse).

KNOWN: ~50% REFUTED on this run -- the convergence gate accepts a slowly-CLIMBING
trim (low spread but drifting carrier); bleeding a drifting trim leaves a residual.
Next: tighten the gate to require a FLAT trim (near-zero slope), not just low spread,
before re-arming. Disarmed to shadow-only meanwhile. [[carrier-loop-absorbing-state]]

+++ lib/stages/gnss/cudaGnssTrack.hpp
@@ -59,13 +59,15 @@ public:
     std::vector<double> dop, cp, cp_rate, dop_rate, ctrim;
     std::vector<long long> ref_hop;
     std::vector<uint8_t> active;
-    /// f_ref TRIM-BLEED (phase 1): a ONE-SHOT per-PRN re-pin request from the broker. When the
-    /// broker bleeds a converged, coherent, standing carrier trim it zeroes its own ctrim and
-    /// sets this so the tracker re-adopts the seed (f_ref = dop) phase-continuously (reanchored
-    /// == 2) -- folding the frozen sub-fence pin offset into f_ref so the despread replica runs
-    /// on-true. Set in set_seeds_callback, CONSUMED+CLEARED at the per-frame snapshot (one re-pin
-    /// per request, never a standing forced re-pin). See gps_distributed_broker --carrier-bleed.
-    std::vector<uint8_t> carrier_repin;
+    /// f_ref TRIM-BLEED (phase 1): a ONE-SHOT per-PRN re-pin AMOUNT (Hz) from the broker. When
+    /// the broker bleeds a converged, coherent, standing trim it zeroes its own ctrim and sends
+    /// the absorbed amount here; the tracker does f_ref += amount (NOT f_ref = dop -- dop is the
+    /// FLEET model, off from this sat's true carrier by exactly the per-sat trim). Adding the
+    /// trim to f_ref leaves the combined carrier f_ref+ctrim INVARIANT (coherence preserved),
+    /// phase-continuous via reanchored==2, and moves the frozen offset out of the despread ramp.
+    /// Set in set_seeds_callback, CONSUMED+CLEARED at the per-frame snapshot (one re-pin per
+    /// request). 0 = no request. See gps_distributed_broker --carrier-bleed.
+    std::vector<double> carrier_repin;
 
     // Tracking control state (pass-1 owns it; command instances execute in frame order).
     std::vector<double> f_ref;         ///< pinned replica carrier per PRN (NaN = unset)
